+++ src/commands/slashCommands/utilities/loadout.js
@@ -6,7 +6,10 @@ module.exports = {
     data: new SlashCommandBuilder()
         .setName('loadout')
         .setDMPermission(true)
-        .setDescription('Picks a random loadout to use in-game.'),
+        .setDescription('Picks a random loadout to use in-game.')
+        .setDescriptionLocalizations({
+            nl: 'Kiest een willekeurige loadout om in-game te gebruiken.'
+        }),
 
     async execute(interaction, auth, langOpt) {