+++ src/commands/slashCommands/utilities/serverStatus.js
@@ -8,7 +8,10 @@ module.exports = {
     data: new SlashCommandBuilder()
         .setName('status')
         .setDMPermission(true)
-        .setDescription('Shows current in-game server status.'),
+        .setDescription('Shows current in-game server status.')
+        .setDescriptionLocalizations({
+            nl: 'Toont huidige in-game server status.'
+        }),
 
     async execute(interaction, auth, langOpt) {